+++ CHANGES.rst
@@ -4,7 +4,8 @@
 
 0.5.7-dev
 =========
- * ``taxit update_taxids --taxid-column`` allows updating of any tax_id column
+ * ``taxit update_taxids --taxid-column`` allows updating of any tax_id column [GH-84]
+ * ``taxit update_taxids --ignore-unknowns`` allows unknown tax_ids to remain in final output [GH-84]
 
 0.5.7
 =====